requirements.txt
streamlit
google-generativeai
Pillow
pdfkit
python-dotenv
torch
diffusers
python-dotenv